backend_old/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic[email]
python-jose[cryptography]
python-multipart
python-dotenv
aiosmtplib
fastapi-mail
slowapi
requests
pytest
httpx